pyproject.toml
[project]
name = "ontology-search"
version = "0.1.0"
description = ""
authors = [{ name = "Olivier Banville", email = "olivier@banville.net" }, { name = "Steven Mugisha Mizero", email = "smugisha@uoguelph.ca" }]
readme = "README.md"
requires-python = "^3.10"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
django = "^5.1.6"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
djangorestframework-simplejwt = "^5.4.0"
psycopg2-binary = "^2.9.10"              # for postgres support
gunicorn = "^23.0.0"                     # for production server
redis = "^5.2.1"
django-redis = "^5.4.0"
django-cors-headers = "^4.7.0"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
whitenoise = "^6.9.0"
rdflib = "^7.1.4"
requests = "^2.32.4"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings.local"
python_files = ["tests.py", "test_*.py", "*_test.py", "testing/python/*.py"]
